+++ src/main/java/fr/miage/agents/api/model/Produit.java
@@ -6,7 +6,7 @@ import java.io.Serializable;
  * Created by Guillaume on 07/11/2016.
  */
 public class Produit implements Serializable {
-    public int idProduit;
+    public long idProduit;
 
     public String nomProduit;
